Make bundleable
So the last pull request was unfinished, sorry that wasn’t communicated.

This bring a bunch of changes to development and how it’s maintained.
Also the `bundle exec localise` has been tested to work.

I’ve moved away from `lokalise-pull` and back to `lokalise`, I wanted
to keep it as `lokalise-pull` so that it would leave room in the ruby
gems naming system. Which at some point could be taken to make a fully
fledged gem with download, upload and more. But that is not a problem
we have right now so we’ll deal with it later.

This also offers exposure to control how the gem structure of the
output files. This enables the option of more output configurations
like usage for ruby-motion.