Implement yaml_include_root flag for Lokalise
Lokalise made a breaking change to their API to not include the
country ISO code as the root key by default. This caused unusable
translation files for us.

This change adds a flag to set the argument for yaml_include_root so
that it can still be included in the response when needed.